Add filters for cart product items
This will allow to modify product values that are used to calculate the
shipping for this product.

An example of supported values are width, height, and length, among
others.

Supports the "Use FlagShip Packing API to pack items into boxes" Box
Split option and may not be supported for remaining methods.